Rename server.jsx > renderMiddleware
